+++ src/config/StaticReflection.ts
@@ -1,0 +1,5 @@
+export enum StaticReflection {
+
+    NONE = 0,
+    CLASS_CONSTRUCTOR = 1
+}