+++ src/target/lua/LuaKeywords.ts
@@ -7,5 +7,6 @@ export enum LuaKeywords {
     CLASS_INIT_FUNCTION_NAME = "__init",
     CLASS_SUPER_REFERENCE_NAME = "__super",
     CLASS_INSTANCE_LOCAL_NAME = "instance",
+    CLASS_STATIC_REFLECTION = "__staticReflection",
     EXPORT_LOCAL_NAME = "__export"
 }